Move class_property module into stpipe
