+++ jwst/stpipe/class_property.py
